# Use array to buffer decoded query instead of ByteBuffer.