@anuraaga 2019-12-16
Forgot about Netty's magical `allocateUninitializedArray` :)