@fenik17 2019-12-16
@anuraaga 
>but I can't find any motivation for that in the PR

That was be done to avoid allocate `CharsetDecoder` on each encoded character - it happens under the hood of `String` constructor. But this moment optimized in the new Java versions for ISO-8859-1, US-ASCII and UTF-8 charsets. You can try banchmarking with non default charset.