@anuraaga 2019-12-16
Thanks for the explanation that helps a lot! I'm only optimizing the utf8 case here (I think it should be by far the most common) so seems like it should be fine.