@normanmaurer 2019-12-16
@netty-bot test this please