@normanmaurer 2019-12-16
@anuraaga @fenik17 we could also just check the `Charset` and depending on it use the `byte[]` or the other code-path... WDYT ?