@anuraaga 2019-12-16
Sorry I mixed this up with my other PR for the encoding... this one is not only touching utf8. Let me check performance of other encodings and then possibly make that branch