@anuraaga 2019-12-17
I added a Shift-JIS benchmark, and performance looks fine (a tiny bit faster than `onlyDecoding` since it's a tiny bit smaller when encoded). I am testing on Java 13 - I checked Java 8 too and see that `CharsetDecoder` is only allocated once for the non-standard charsets, while on Java 9+ it adds the even more optimized codepaths for the standard ones. Didn't go to Java 7 but I'm assuming the issue was fixed in Java 8.

The implementation for decoding non-standard is actually wrapping the passed in array with `ByteBuffer` and doing the same as we had before. Technically there's a small amount of extra garbage because of the wrapping, but I don't think it's significant enough to create a special codepath just for non-standard and suggest going as is.

```
Benchmark                                         Mode  Cnt  Score   Error   Units
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1.407 ± 0.134  ops/us
```