@normanmaurer 2019-12-18
@netty-bot test this please